2-D array

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -25,7 +25,7 @@ int **alloc_grid(int width, int height)
 	/* allocate memory for columns */
 	for (i = 0; i < height; i++)
 	{
-		grid[i] = (int) malloc(width * sizeof(int));
+		grid[i] = malloc(width * sizeof(int));
 
 		/* check for NULL and roll back */
 		if (grid[i] == NULL)
